+++ inbound.go
@@ -11,8 +11,8 @@ import (
 	"github.com/kalifun/vda5050-types-go/state"
 	"github.com/kalifun/vda5050-types-go/visualization"
 
-	"github.com/kalifun/navlink/gerrors"
-	"github.com/kalifun/navlink/session"
+	"github.com/kalifun/navlink/internal/gerrors"
+	"github.com/kalifun/navlink/internal/session"
 	"github.com/kalifun/navlink/topic"
 )