+++ internal/mqtt/wait.go
@@ -6,7 +6,7 @@ import (
 
 	pahomqtt "github.com/eclipse/paho.mqtt.golang"
 
-	"github.com/kalifun/navlink/gerrors"
+	"github.com/kalifun/navlink/internal/gerrors"
 )
 
 const defaultTokenWait = 30 * time.Second